requirements.txt
torch>=2.2
torchaudio>=2.2
pretty_midi
mido
numpy
scipy
tqdm
pyyaml
music21
soundfile
librosa
transformers>=4.41
datasets>=2.20
diffusers>=0.30
accelerate>=0.33